+++ platforms/gradle/foundry-gradle-plugin/src/main/kotlin/foundry/gradle/FoundryProperties.kt
@@ -409,10 +409,6 @@ internal constructor(
   public val moshixGenerateProguardRules: Boolean
     get() = booleanProperty("moshix.generateProguardRules", defaultValue = true)
 
-  /** Flag to connect SqlDelight sources to KSP. */
-  public val kspConnectSqlDelight: Boolean
-    get() = booleanProperty("foundry.ksp.connect.sqldelight")
-
   /** Flag to connect ViewBinding sources to KSP. */
   public val kspConnectViewBinding: Boolean
     get() = booleanProperty("foundry.ksp.connect.viewbinding")
@@ -812,6 +808,18 @@ internal constructor(
         .optionalStringProvider("foundry.jvm.target")
         .mapToInt()
         .orElse(resolver.gradleProviderOf(versions.jvmTarget))
+
+  /**
+   * Controls the jvm-default behavior. Unlike KGP, we default to `NO_COMPATIBILITY`.
+   *
+   * @see <a
+   *   href="Docs">https://kotlinlang.org/docs/whatsnew22.html#changes-to-default-method-generation-for-interface-functions</a>
+   */
+  public val kotlinJvmDefaultMode: Provider<String>
+    get() =
+      resolver
+        .optionalStringProvider("foundry.kotlin.jvm.jvmDefaultMode")
+        .orElse("no_compatibility")
 
   /** Flag to disable JVM vendor setting locally. */
   public val jvmVendorOptOut: Boolean